pyproject.toml
[project]
name = "staged-rag-mcp"
version = "0.1.0"
description = "MCP-based staged retrieval augmented generation server."
readme = "README.md"
authors = [
    { name = "Shashidhar Reddy Nalamari", email = "nalamarishashidharreddy@gmail.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.5.0",
    "pydantic>=2.5",
    "numpy>=1.26",
    "rank-bm25>=0.2.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "google-genai>=1.0",
    "pypdf>=4.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov",
    "mypy"
]
faiss = [
    "faiss-cpu>=1.8.0"
]
openai = [
    "openai>=1.0"
]
ollama = [
    "ollama>=0.2"
]
huggingface = [
    "sentence-transformers>=2.0"
]
together = [
    "together>=1.0"
]
all-providers = [
    "openai>=1.0",
    "ollama>=0.2",
    "sentence-transformers>=2.0",
    "together>=1.0"
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
